Skip some tests if KO_DOCKER_REPO is missing and explicitely asked for…
… instead of "hard" failing, we could just skip it as it doesn't meet
the requirements. Impacted tests :

- TestHelmDeployPipelineRun
- TestKanikoTaskRun

To skip those tests instead of failing, you need to provide the
following arguments to go test command :
`-ldflags '-X github.com/tektoncd/pipeline/test.missingKoFatal=false'`
